Changes to be committed: 	modified:   src/shknam.c 	modified:   src/trap.c 	modified:   src/zap.c
While polymorph was clearing the name because poly_obj
actually creates a brand new obj and copies field values
over, water_damage does not so the name persisted on
the blank spellbook. In this case, that isn't really
appropriate. Clear the name.

Ensure that cancellation has no effect on these
ordinary books.

Make the second-hand bookstore an option for holding
the book tribute item too.